+++ include/magicxx/magic_exception.hpp
@@ -73,6 +73,25 @@ public:
 };
 
 /**
+ * @class path_is_not_directory
+ *
+ * @brief Exception thrown from magic::identify_files(directory)
+ *        when the path is not a directory.
+ */
+class path_is_not_directory final : public magic_exception {
+public:
+    /**
+     * @brief Construct path_is_not_directory with
+     *        the path that is not a directory.
+     *
+     * @param[in] path                  The path that is not a regular file.
+     */
+    explicit path_is_not_directory(const std::string& path)
+      : magic_exception{std::format("'{}' is not a directory.", path)}
+    { }
+};
+
+/**
  * @class path_does_not_exist
  *
  * @brief Exception thrown when a path does not exist.
@@ -86,6 +105,24 @@ public:
      */
     explicit path_does_not_exist(const std::string& path)
       : magic_exception{std::format("'{}' does not exist.", path)}
+    { }
+};
+
+/**
+ * @class filesystem_error
+ *
+ * @brief Exception thrown when the underlying std::filesystem OS API fails.
+ */
+class filesystem_error final : public magic_exception {
+public:
+    /**
+     * @brief Construct filesystem_error with a path and an error message.
+     *
+     * @param[in] path                  The path related to the error.
+     * @param[in] error_message         The description of the error.
+     */
+    filesystem_error(const std::string& path, const std::string& error_message)
+      : magic_exception{std::format("'{}': {}.", path, error_message)}
     { }
 };
 
@@ -122,25 +159,25 @@ public:
 };
 
 /**
- * @class magic_load_error
+ * @class magic_load_database_file_error
  *
  * @brief Exception thrown when magic::load_database_file fails.
  */
-class magic_load_error final : public magic_exception {
+class magic_load_database_file_error final : public magic_exception {
 public:
     /**
-     * @brief Construct magic_load_error with an error message
+     * @brief Construct magic_load_database_file_error with an error message
      *        and the path of the database file.
      *
      * @param[in] error_message         The description of the error.
      * @param[in] database_file_path    The path of the magic database file.
      */
-    magic_load_error(
+    magic_load_database_file_error(
         const std::string& error_message,
         const std::string& database_file_path
     )
       : magic_exception{
-            std::format("magic_load({})", database_file_path),
+            std::format("magic_load_database_file({})", database_file_path),
             error_message
         }
     { }